store pointsizes in ints instead of floats - apparently a customer hit the 3276 point maximum (and didn't like it)... this is binary compatible - we are only changing internal data structures
[git-p4: depot-paths = "//depot/qt/main/": change = 38792]

+++ src/kernel/qfontdata_p.h
@@ -78,8 +78,8 @@ struct QFontDef {
 
     QString family;
 
-    short pixelSize;
-    short pointSize;
+    int pixelSize;
+    int pointSize;
     short lbearing;
     short rbearing;
